+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 ## Unreleased
 
 ### Fixes
-
 - Send UI Profiling app start chunk when it finishes ([#4423](https://github.com/getsentry/sentry-java/pull/4423))
+- Republish Javadoc [#4457](https://github.com/getsentry/sentry-java/pull/4457)
 
 ## 8.13.2